DataValidation abstraction separation
