+++ src/main/java/demo/Logic.java
@@ -7,18 +7,6 @@ public class Logic {
     private List<Integer> numbers;
     private String source;
 
-    public static boolean validate(String input) {
-        if (input == null)
-            return false;
-
-        for (int i = 0; i < input.length(); i++) {
-            var currentCh = input.charAt(i);
-            if (currentCh < '0' || currentCh > '9')
-                return false;
-        }
-        return true;
-    }
-
     public Logic(String number) {
         if (number == null)
             throw new NullPointerException();
